Edited confusing usage example in README
The android:textColorHint attribute of an EditText does not have any visible effect when used within a MaterialTextField. Viewers of the usage example may be confused as to why the attribute does not work upon use in applications. As such, I removed this line for clarity and added an app:mtf_labelColor attribute to its parent view.